online normalizer

+++ espnet2/layers/fdlp_spectrogram.py
@@ -452,7 +452,7 @@ class fdlp_spectrogram(torch.nn.Module):
 
         ## Adjust the phase
         phi = (phase[-1] - phase[0]) / phase.shape[0]
-        x_ph = torch.arange(phase.shape[0], dtype=frames.dtype, device=frames.device)
+        x_ph = torch.arange(phase.shape[0], device=frames.device)
         y_ph = phase[0] + x_ph * phi
         ph_corrected = y_ph - phase
         ph_corrected = ph_corrected * phase_max_cap / np.max(ph_corrected)
